% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Power_Analysis_ShortTermSFE_SingleIntervention.R
\name{sim.power.shortsfe.sinint}
\alias{sim.power.shortsfe.sinint}
\title{Estimate Empirical Power for Short-Term Semi-Field Experiment Testing Single Intervention}
\usage{
sim.power.shortsfe.sinint(
  n.ch.per.trt,
  lambda,
  intvn.effect,
  chamber.var,
  nsim,
  n.cores = 1,
  use.random = TRUE
)
}
\arguments{
\item{n.ch.per.trt}{An integer specifying the number of chambers allocated per treatment group.}

\item{lambda}{A numeric value indicating the expected mean mosquito count in control chambers.}

\item{intvn.effect}{A numeric value representing the proportional reduction in mosquito count due to the intervention (e.g., ITN).}

\item{chamber.var}{A numeric value specifying the variance of random chamber-level effects.}

\item{nsim}{An integer indicating the total number of simulations.}

\item{n.cores}{An integer the number of cores to use for parallel processing.}

\item{use.random}{A logical value indicating whether to plot mosquito counts simulated using a sampling distribution (with random or fixed effects).
If \code{TRUE}, returns power for expected mosquito counts simulated using sampling distribution with random effects;
If \code{FALSE}, returns power for expected counts using sampling distribution based on fixed effects only.}
}
\value{
A named numeric vector:
\describe{
  \item{power}{Estimated empirical power (rounded to two decimal places)}
  \item{ci.lower}{Lower bound of 95\% confidence interval}
  \item{ci.upper}{Upper bound of 95\% confidence interval}
}
}
\description{
Runs repeated simulations and Poisson GLMM fits to estimate empirical power
as the proportion of simulations with p-values below 0.05. These p-values are generated using the function called `sim.pval.shortsfe.sinint()`.
}
\note{
Parallel execution is supported via `n.cores`, but examples default to `n.cores = 1` for reproducibility and package checks.
}
\examples{
# For a realistic SFE design, nsim should be at least 1000.

sim.power.shortsfe.sinint(
  n.ch.per.trt = 2,
  lambda = 50,
  intvn.effect = 0.8,
  chamber.var = 0.1807,
  nsim = 2,
  n.cores = 1,
  use.random = TRUE
)

}
