% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpowa.R
\docType{package}
\name{simpowa-package}
\alias{simpowa}
\alias{simpowa-package}
\title{Power Analysis and Sample Size Calculation for Semi-Field Vector Control Trials}
\description{
Uses simulations from generalized linear mixed-effects models to incorporate random effects across multiple sources and levels of variation, and a dispersion parameter to account for overdispersion and capture unexplained variability.
Covers design scenarios for both short-term and long-term trials evaluating the impact of single or combined vector control interventions.
Methods build on Kipingu et al. (2025) <doi:10.1186/s12936-025-05454-y> and Johnson et al. (2015) <doi:10.1111/2041-210X.12306>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/akipingu/simpowa}
  \item Report bugs at \url{https://github.com/akipingu/simpowa/issues}
}

}
\author{
\strong{Maintainer}: Andrea M. Kipingu \email{akipingu@ihi.or.tz} (\href{https://orcid.org/0000-0002-9063-8651}{ORCID}) (Environmental Health & Ecological Sciences Department, Ifakara Health Institute, P.O. Box 78 373, Dar es Salaam, Tanzania)

Authors:
\itemize{
  \item Mafalda Viana \email{mafalda.viana@glasgow.ac.uk} (\href{https://orcid.org/0000-0001-5975-6505}{ORCID}) (School of Biodiversity, One Health and Veterinary Medicine, University of Glasgow, Graham Kerr Building, Glasgow, G12 8QQ, UK)
  \item Paul C.D. Johnson \email{paul.johnson@glasgow.ac.uk} (\href{https://orcid.org/0000-0001-6663-7520}{ORCID}) (School of Biodiversity, One Health & Veterinary Medicine, University of Glasgow, Wolfson Link Building, Glasgow, G12 8QQ, UK)
  \item Samson S. Kiware \email{skiware@ihi.or.tz} (\href{https://orcid.org/0000-0001-7252-520X}{ORCID}) (Environmental Health & Ecological Sciences Department, Ifakara Health Institute, P.O. Box 78 373, Dar es Salaam, Tanzania)
}

}
