% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpowa.R
\docType{package}
\name{simpowa-package}
\alias{simpowa}
\alias{simpowa-package}
\title{simpowa: A Generalizable Simulation-Based Power Analysis R Package for Semi-Field Vector Control Trials}
\description{
The methods for a simulation-based power analysis tailored to short- or long-term semi-field vector control trials testing single or combined interventions.
The \code{simpowa} package supports random effects modeling, empirical p-value extraction, and
parallelized workflows. It is designed to guide experimental design and improve statistical rigor
in entomological research.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/akipingu/simpowa}
  \item Report bugs at \url{https://github.com/akipingu/simpowa/issues}
}

}
\author{
\strong{Maintainer}: Andrea M. Kipingu \email{akipingu@ihi.or.tz} (\href{https://orcid.org/0000-0002-9063-8651}{ORCID}) (Department of Environmental Health & Ecological Sciences, Ifakara Health Institute, P.O. Box 78 373, Dar es Salaam, Tanzania)

Authors:
\itemize{
  \item Mafalda. Viana \email{mafalda.viana@glasgow.ac.uk} (\href{https://orcid.org/0000-0001-5975-6505}{ORCID}) (School of Biodiversity, One Health and Veterinary Medicine, University of Glasgow, Graham Kerr Building, Glasgow, G12 8QQ, UK)
  \item Paul C.D. Johnson \email{paul.johnson@glasgow.ac.uk} (\href{https://orcid.org/0000-0001-6663-7520}{ORCID}) (School of Biodiversity, One Health & Veterinary Medicine, University of Glasgow, Wolfson Link Building, Glasgow, G12 8QQ, UK)
  \item Samson S. Kiware \email{skiware@ihi.or.tz} (\href{https://orcid.org/0000-0001-7252-520X}{ORCID}) (Department of Environmental Health & Ecological Sciences, Ifakara Health Institute, P.O. Box 78 373, Dar es Salaam, Tanzania)
}

}
